requirements.txt
autogen-core==0.4.0.dev6
autogen-ext[openai]==0.4.0.dev6
python-dotenv==1.0.1
chainlit==1.3.1
pydantic==2.9.2
opentelemetry-api==1.27.0
python-decouple==3.8
